+++ packages/ui/src/components/molecules/index.ts
@@ -11,8 +11,15 @@ export {
   type TCheckboxFilterWidget,
 } from './CheckboxFilterWidget'
 export { CreateAuthor, type TCreateAuthor } from './CreateAuthor'
+export {
+  EditableItem,
+  type TItem,
+  type TItemAvailable,
+  type TItemRow,
+} from './EditableItem'
 export { InputGroup, type TInputGroup } from './InputGroup'
 export { Modal, type TModal } from './Modal'
+export { ResourceTitleLink, type TResourceTitleLink } from './ResourceTitleLink'
 export { SelectGroup, type TSelectGroup } from './SelectGroup'
 export { Tabs, type TTabs } from './Tabs'
 export { TextareaGroup, type TTextareaGroup } from './TextareaGroup'